Add files via upload
Change Mentor to extend User class

Updated constructors to reflect that.

Fixed User and Feedback objects not being comparable

Made User and Feedback objects comparable to instances of
themselves to correct JUnit tests that used these classes.

Add createMockup method for User, Mentor and Course for debugging

Added createMockup method that creates an instance of the
class with mockup values for attributes. They are to be used for
debugging only.

Change CoursePost and FeedbackPanel parameters

Replaced CoursePost and FeedbackPanel constructors
parameters from TODO to the respective classes they use.